// app/modules/offer/store/offer-comments/offer-comment.action-types.ts
export enum OfferCommentsActionEnums {
  Get_All_Offer_Comments = '[Offer Comments] Get All Comments',
  Get_All_Offer_Comments_Success = '[Offer Comments] Get Comment Success',
  Load_More_Offer_Comments = '[Offer Comments] Lode More Comment',
  Load_More_Offer_Comments_Success = '[Offer Comments] Lode More Comment Success',
  Load_More_Offer_Comments_Reply = '[Offer Comments] Lode More Comment Reply',
  Load_More_Offer_Comments_Reply_Success = '[Offer Comments] Lode More Comment Reply Success',
  Get_All_Offer_Comments_Failure = '[Offer Comments] Get Comment Failure',
  Add_Offer_Comment = '[Offer Comments] Add comment',
  Add_Offer_Comment_Reply = '[Offer Comments] Add comment Reply',
  Edit_Offer_Comment = '[Offer Comments] Edit comment',
  Edit_Offer_Comment_Reply = '[Offer Comments] Edit comment Reply',
  Delete_Offer_Comment = '[Offer Comments] Delete comment',
  Delete_Offer_Comment_Reply = '[Offer Comments] Delete comment Reply',
}

export enum OfferAuctionActionEnums {
  OFFER_AUCTION_CLOSE = '[Offer Auction] Close',
  OFFER_AUCTION_BUY = '[Offer Auction] Buy',
  OFFER_AUCTION_SUCCESS = '[Offer Auction] Success',
}

export enum CURRENT_AUCTION_MODAL {
  AUCTION_CLOSE = 'close-auction-modal',
  AUCTION_BUY = 'buy-auction-modal',
  AUCTION_SUCCESS = 'success-auction-modal',  
}